Fixed rendering GuiButtonCategory on landing page.
