--- Xplat/src/main/java/vazkii/patchouli/client/book/page/DummySimpleInventory.java
@@ -1,10 +1,0 @@
-package vazkii.patchouli.client.book.page;
-
-import net.minecraft.world.Container;
-import net.minecraft.world.SimpleContainer;
-
-public final class DummySimpleInventory {
-    public static final Container INSTANCE = new SimpleContainer(1);
-
-    private DummySimpleInventory() {}
-}
