+++ Xplat/src/main/java/vazkii/patchouli/client/book/page/PageCrafting.java
@@ -4,6 +4,9 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.player.StackedItemContents;
+import net.minecraft.world.inventory.CraftingContainer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.level.Level;
@@ -12,6 +15,8 @@ import vazkii.patchouli.client.book.page.abstr.PageDoubleRecipeRegistry;
 import vazkii.patchouli.common.util.RecipeUtil;
 
 import java.util.*;
+
+import org.jetbrains.annotations.NotNull;
 
 public class PageCrafting extends PageDoubleRecipeRegistry<CraftingRecipe> {
 
@@ -73,33 +78,89 @@ public class PageCrafting extends PageDoubleRecipeRegistry<CraftingRecipe> {
 			parent.renderItemStack(graphics, recipeX + 79, recipeY + 41, mouseX, mouseY, workstation);
 		}
 	}
-	public static final class WorkstationIcons {
 
-		public static final Map<RecipeType<?>, ItemStack> ICONS = Map.of(
-				RecipeType.CRAFTING, new ItemStack(net.minecraft.world.item.Items.CRAFTING_TABLE),
-				RecipeType.SMELTING, new ItemStack(net.minecraft.world.item.Items.FURNACE),
-				RecipeType.BLASTING, new ItemStack(net.minecraft.world.item.Items.BLAST_FURNACE),
-				RecipeType.SMOKING, new ItemStack(net.minecraft.world.item.Items.SMOKER),
-				RecipeType.CAMPFIRE_COOKING, new ItemStack(net.minecraft.world.item.Items.CAMPFIRE),
-				RecipeType.STONECUTTING, new ItemStack(net.minecraft.world.item.Items.STONECUTTER),
-				RecipeType.SMITHING, new ItemStack(net.minecraft.world.item.Items.SMITHING_TABLE)
-
-		);
-
-		public static ItemStack iconFor(RecipeType<?> type) {
-			return ICONS.getOrDefault(type, ItemStack.EMPTY);
-		}
-	}
 
 	@Override
 	protected int getRecipeHeight() {
 		return 78;
 	}
 
+	
 	@Override
 	public ItemStack getRecipeOutput(CraftingRecipe recipe) {
 		var regs = RecipeUtil.getRegistryAccess().orElse(null);
 		if (regs == null) return ItemStack.EMPTY;
 		return recipe.assemble(DummyCraftingInventory.INSTANCE.asCraftInput(), regs);
+	}
+
+	public static final class DummyCraftingInventory implements CraftingContainer {
+		public static final DummyCraftingInventory INSTANCE = new DummyCraftingInventory();
+
+		public DummyCraftingInventory() {}
+
+		@Override
+		public int getWidth() {
+			return 3;
+		}
+
+		@Override
+		public int getHeight() {
+			return 3;
+		}
+
+		@Override
+		public net.minecraft.core.@NotNull NonNullList<ItemStack> getItems() {
+			return net.minecraft.core.NonNullList.withSize(getContainerSize(), ItemStack.EMPTY);
+		}
+
+		@Override
+		public int getContainerSize() {
+			return getWidth() * getHeight();
+		}
+
+		@Override
+		public boolean isEmpty() {
+			return true;
+		}
+
+		@Override
+		public @NotNull ItemStack getItem(int i) {
+			return ItemStack.EMPTY;
+		}
+
+		@Override
+		public @NotNull ItemStack removeItem(int i, int j) {
+			return ItemStack.EMPTY;
+		}
+
+		@Override
+		public @NotNull ItemStack removeItemNoUpdate(int i) {
+			return ItemStack.EMPTY;
+		}
+
+		@Override
+		public void setItem(int i, @NotNull ItemStack itemStack) {
+			// NO-OP
+		}
+
+		@Override
+		public void setChanged() {
+			// NO-OP
+		}
+
+		@Override
+		public boolean stillValid(@NotNull Player player) {
+			return false;
+		}
+
+		@Override
+		public void clearContent() {
+			// NO-OP
+		}
+
+		@Override
+		public void fillStackedContents(@NotNull StackedItemContents stackedItemContents) {
+			// NO-OP
+		}
 	}
 }
