--- Xplat/src_main/resources/assets/patchouli/items/model/book_gray.json
@@ -1,3 +1,0 @@
-{
-    "parent": "patchouli:models/item/book_gray"
-}
