+++ run/patchouli_books/cookbook/en_us/entries/advanced_meals/shapeless_test.json
@@ -1,6 +1,6 @@
 {
   "name": "Shapeless Test",
-  "icon": "minecraft:",
+  "icon": "minecraft:stone",
   "category": "patchouli:advanced_meals",
   "pages": [
     {
